pyproject.toml
[tool.poetry]
name = "learn-fastapi-user-service"
version = "0.1.0"
description = ""
authors = ["sky <sky@sky.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
tortoise-orm = "^0.20.0"
bcrypt = "^4.0.1"
fastapi = "^0.104.1"
jose = "^1.0.0"
uvicorn = "^0.23.2"
ruff = "^0.1.3"
aerich = "^0.7.2"
pyjwt = "^2.8.0"
pytest-asyncio = "^0.21.1"
pytest = "^7.4.3"
ipdb = "^0.13.13"


[tool.aerich]
tortoise_orm = "db.tortoise_connection.TORTOISE_ORM"
location = "./src/db/migrations"
src_folder = "./src/"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
